ai_career_match-master/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
#numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
python-multipart==0.0.12
pydantic==2.10.0
streamlit==1.40.0
pytest==8.3.3
#cv
nltk==3.8.1
# DEPRECATED: Local CV parsing disabled - using ResumeParser.app API instead
# pypdf2==3.0.1
# docx2txt==0.8
beautifulsoup4==4.12.2
requests==2.31.0
#httpx==0.27.0
ollama==0.1.7
nltk==3.8.1
sentence-transformers==2.2.2
google-generativeai==0.3.2
Jinja2==3.1.6
numpy==2.3.0
pillow==11.2.1
PyMuPDF==1.26.1
python-docx==1.1.2
RapidFuzz==3.13.0
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
#streamlit==1.45.1